# Optimized indexOf to speed up framing stage