@johanandren 2016-09-23
Among other things, avoiding allocation of a closure for each call (supertype called `indexWhere(_ == elem, from)`)