@johanandren 2016-09-23
Note that the bench names are a bit misleading as I at first expected the `ByteStrings.indexOf` to be the problem, but it turned out `ByteString1.indexOf` was a large part of it.