@monkey-mas 2016-09-25
>Note that the bench names are a bit misleading as I at first expected the ByteStrings.indexOf to be the problem, but it turned out ByteString1.indexOf was a large part of it.

I suspect the real problem we have here is `CompactByteString` but not `ByteString1` :) I guess we should implement `CompactByteString#indexOf` as `compact` returns `CompactByteString`. We need the method not to call superclass' `indexOf(...)`?

`CompactByteString#indexOf` can be:
```
override def indexOf[B >: Byte](elem: B, from: Int): Int =
  if (from >= length) -1
  else toByteString1.indexOf(elem, from)
```