@johanandren 2016-10-18
Thought I needed to verify this but now I have seen it take down one of the FramingSpec tests from 20s to 2s so, definitely an improvement for framing.