@ktoso 2016-10-18
These look important and good enough to me to include. 
The tests against Vector can be implemented in a different ticket perhaps? Since it's the Scala collection that does weird things.

LGTM