@johanandren 2016-10-18
I fixed those instead. Was annoyed to have disabled them and left like that. - also found a bug or two when I re-enabled them so good thing I did.