@ktoso 2016-10-18
Even better! LGTM, good catch in there. I think it's worth pushing the perf on these operations.