@ktoso 2016-10-19
@johanandren LGTM, but please squash manually and then merge - via github UI it will end up attributing the change to you instead of @monkey-mas 

Thanks @monkey-mas !