@johanandren 2016-10-19
@ktoso it actually is my change, @monkey-mas just reviewed it