@ktoso 2016-10-19
Whoops, sorry :) Confused it with the other perf opts that came in as a batch from @monkey-mas :-)
Thanks @johanandren !